Completed battle menu

+++ src/main/java/despacito7/detail/Monster.java
@@ -48,9 +48,9 @@ public class Monster extends AnimatingObject implements Cloneable {
         return id;
     }
 
-    public int getStat(Stat s) {
+    public float getStat(Stat s) {
         System.out.println(s + ": " + stats.get(s));
-        return (int) stats.get(s);
+        return stats.get(s).floatValue();
     }
 
     public ArrayList<Move> getMoves() {
